複数のFlagを生成する問題を修正
