+++ src/knowledge/respawn_record.ts
@@ -1,0 +1,7 @@
+/**
+ * -
+ *   - 20000 ticks 7 extensions 1 tower and RCL3 658/135k
+ * - v6.6.48
+ *   - 6500 to RCL3
+ *   - 6900 1st tower
+ */
